Changed Overview and details to be child fragments instead of seperate fragments
